pyproject.toml
[tool.poetry]
name = "python-bot"
version = "0.1.0"
description = ""
authors = ["Alexandr Ovsyannikov <alexov2004@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
dill = "^0.3.7"
aiogram = "2.25.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-random = "^0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
